修复一些小bug，目前实现步兵稳定跟随前哨站

+++ io/command.hpp
@@ -7,8 +7,8 @@ struct Command
 {
   bool control;
   bool shoot;
-  double yaw;
-  double pitch;
+  float yaw;
+  float pitch;
   double horizon_distance = 0;  //无人机专有
 };
